requirements.txt
# Core ML
tensorflow>=2.4

# Image processing
opencv-python-headless>=4.5.2,<4.6
scikit-image>=0.17
Pillow>=7.2

# Data augmentation
albumentations>=1.0

# Utilities
numpy>=1.19
scipy>=1.5
matplotlib>=3.3
imutils>=0.5
natsort>=7.1
requests>=2.24
scikit-learn>=0.24

plotly>=6.5.0
nbformat>=5.10.4